Mini fixs

+++ client/src/widgets/refine/ImagesPreview.tsx
@@ -1,6 +1,15 @@
 "use client";
 
-import { Box, IconButton } from "@mui/material";
+import {
+  Box,
+  Button,
+  Dialog,
+  DialogActions,
+  DialogContent,
+  DialogContentText,
+  DialogTitle,
+  IconButton,
+} from "@mui/material";
 import { Close } from "@mui/icons-material";
 import { useDelete } from "@refinedev/core";
 import Image from "next/image";
@@ -20,22 +29,36 @@ export const ImagesPreview = ({
   const [savedImages, setSavedImages] = useState<string[]>(
     (images as string[]) || [],
   );
+  const [open, setOpen] = useState(false);
+  const [imageToDelete, setImageToDelete] = useState<string | null>(null);
 
-  const handleDelete = (imageWillDelete: string) => {
-    if (!id) return;
+  const handleClickOpen = (image: string) => {
+    setImageToDelete(image);
+    setOpen(true);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+    setImageToDelete(null);
+  };
+
+  const handleDelete = () => {
+    if (!id || !imageToDelete) return;
+
     deleteImage(
       {
         resource,
         id,
         values: {
-          image: imageWillDelete,
+          image: imageToDelete,
         },
       },
       {
         onSuccess: async () => {
           setSavedImages((prev) =>
-            prev.filter((item) => item !== imageWillDelete),
+            prev.filter((item) => item !== imageToDelete),
           );
+          handleClose();
         },
       },
     );
@@ -44,31 +67,54 @@ export const ImagesPreview = ({
   if (!images) return null;
 
   return (
-    <Box sx={{ display: "flex", gap: 2, flexWrap: "wrap" }}>
-      {savedImages.map((image) => {
-        return (
-          <Box key={image} sx={{ position: "relative" }}>
-            <Image
-              src={"/" + image}
-              alt="portfolio image"
-              width={150}
-              height={150}
-              style={{ objectFit: "cover" }}
-            />
-            <IconButton
-              sx={{
-                position: "absolute",
-                top: 0,
-                right: 0,
-                backgroundColor: "rgba(255, 255, 255, 0.7)",
-              }}
-              onClick={() => handleDelete(image)}
-            >
-              <Close />
-            </IconButton>
-          </Box>
-        );
-      })}
-    </Box>
+    <>
+      <Box sx={{ display: "flex", gap: 2, flexWrap: "wrap" }}>
+        {savedImages.map((image) => {
+          return (
+            <Box key={image} sx={{ position: "relative" }}>
+              <Image
+                src={"/" + image}
+                alt="portfolio image"
+                width={150}
+                height={150}
+                style={{ objectFit: "cover" }}
+              />
+              <IconButton
+                sx={{
+                  position: "absolute",
+                  top: 0,
+                  right: 0,
+                  backgroundColor: "rgba(255, 255, 255, 0.7)",
+                }}
+                onClick={() => handleClickOpen(image)}
+              >
+                <Close />
+              </IconButton>
+            </Box>
+          );
+        })}
+      </Box>
+      <Dialog
+        open={open}
+        onClose={handleClose}
+        aria-labelledby="alert-dialog-title"
+        aria-describedby="alert-dialog-description"
+      >
+        <DialogTitle id="alert-dialog-title">
+          {"Підтвердження видалення"}
+        </DialogTitle>
+        <DialogContent>
+          <DialogContentText id="alert-dialog-description">
+            Ви точно хочете видалити це зображення?
+          </DialogContentText>
+        </DialogContent>
+        <DialogActions>
+          <Button onClick={handleClose}>Скасувати</Button>
+          <Button color="error" onClick={handleDelete} autoFocus>
+            Видалити
+          </Button>
+        </DialogActions>
+      </Dialog>
+    </>
   );
 };
